+++ tests/misc.js
@@ -3,6 +3,8 @@ var test = require('tape')
 var anymatch = require('anymatch')
 var rimraf = require('rimraf')
 var memdb = require('memdb')
+var memdown = require('memdown')
+var hyperdrive = require('hyperdrive')
 var encoding = require('dat-encoding')
 var fs = require('fs')
 var os = require('os')
@@ -92,6 +94,27 @@ test('custom db option', function (t) {
     dat.archive.open(function () {
       // Need open otherwise get DeferredLevelDOWN
       t.ok(dat.db.db instanceof require('memdown'), 'db is memdown')
+      try {
+        fs.statSync(path.join(shareFolder, '.dat'))
+        t.fail('.dat folder exists =(')
+      } catch (e) {
+        t.pass('.dat folder does not exist')
+      }
+      dat.close(function () {
+        t.end()
+      })
+    })
+  })
+})
+
+test('custom drive option', function (t) {
+  rimraf.sync(path.join(shareFolder, '.dat'))
+  var drive = hyperdrive(memdb())
+  Dat(shareFolder, {drive: drive}, function (err, dat) {
+    t.error(err)
+    dat.archive.open(function () {
+      // Need open otherwise get DeferredLevelDOWN
+      t.ok(dat.db.db instanceof memdown, 'db is memdown')
       try {
         fs.statSync(path.join(shareFolder, '.dat'))
         t.fail('.dat folder exists =(')
@@ -225,7 +248,7 @@ test('expose swarm.connected', function (t) {
     var network = shareDat.joinNetwork()
     t.equal(network.connected, 0, '0 peers')
 
-    network.swarm.once('connection', function () {
+    network.once('connection', function () {
       t.ok(network.connected >= 1, '>=1 peer')
 
       downDat.close(function (err) {